+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 ### Added
+- COVID-19 alert for printed reproductions
 - laravel/tinker
 
 ### Fixed